Day_2. HW_6 is ready.
