+++ .gitignore
@@ -1,1 +1,6 @@
-.idea
+# Intellij
+/.idea/
+/target/
+*.iml
+*.iws
+.gitignore
