--- src/FamilyTreeIterator.java
@@ -1,28 +1,0 @@
-package homeWork.src;
-
-import java.util.Iterator;
-import java.util.List;
-
-public class FamilyTreeIterator<T> implements Iterator<T> {
-
-    private List<T> members;
-    private int index;
-
-    public FamilyTreeIterator(List<T> members) {
-        this.members = members;
-    }
-
-    public int getIndex() {
-        return index;
-    }
-
-    @Override
-    public boolean hasNext() {
-        return index < members.size();
-    }
-
-    @Override
-    public T next() {
-        return members.get(index++);
-    }
-}
