--- src/Writable.java
@@ -1,11 +1,0 @@
-package homeWork.src;
-
-import java.io.Serializable;
-
-public interface Writable {
-
-    void save(Serializable serializable);
-
-    Object read();
-
-}
