--- src/commands/Commands.java
@@ -1,8 +1,0 @@
-package homeWork.src.commands;
-
-public interface Commands {
-    //    выполняемая команда
-    void execute();
-    //    описание
-    String description();
-}
