+++ src/familyTree/Writable.java
@@ -1,0 +1,11 @@
+package homeWork.src.familyTree;
+
+import java.io.Serializable;
+
+public interface Writable {
+
+    void save(Serializable serializable);
+
+    Object read();
+
+}
