+++ src/familyTree/commands/Commands.java
@@ -1,0 +1,8 @@
+package homeWork.src.familyTree.commands;
+
+public interface Commands {
+    //    выполняемая команда
+    void execute();
+    //    описание
+    String description();
+}
